lesson-4-func  - Completed homework for lesson 4 on func components
