+++ src/react/veiws/layouts/main.js
@@ -1,0 +1,17 @@
+import React from "react";
+import Header from "./header";
+
+import { Outlet } from "react-router-dom";
+
+const Main = () => {
+    return (
+        <>
+            <Header />
+            <div style={{ position: 'relative' }}>
+                <Outlet />
+            </div>
+        </>
+    );
+};
+
+export default Main;
