+++ src/react/veiws/notFound.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const NotFound = () => {
+    return (
+        <h1>Старница не найдена</h1>
+    );
+};
+
+export default NotFound;
