+++ src/react/veiws/profile.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Profile = () => {
+    return (
+        <h1>Профиль</h1>
+    );
+};
+
+export default Profile;
