+++ src/style/index.scss
@@ -21,3 +21,4 @@ code {
 }
 
 @import "components/index";
+@import "views/layouts/index";
