+++ src/style/views/layouts/header.scss
@@ -1,0 +1,4 @@
+.header {
+  display: flex;
+  justify-content: center;
+}
