+++ src/style/views/layouts/index.scss
@@ -1,0 +1,1 @@
+@import "header";
